% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{data.correction}
\alias{data.correction}
\title{Trim and correct data}
\usage{
data.correction(df, taper.correction, fill_missing)
}
\arguments{
\item{taper.correction}{TRUE or FALSE, are you willing to apply Cushman et al (2014) taper correction?}

\item{fill_missing}{TRUE or FALSE, are you willing to extrapolate missing DBH from surrounding DBH?}
}
\value{
a data.table (data.frame) with all relevant variables.
}
\description{
Stack all censuses together and correct DBH, if required
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
