% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{data.prep}
\alias{data.prep}
\title{CTFS-formated data preparation}
\usage{
data.prep(site, stem, taper.correction, fill.missing, palm, strangler,
  maxrel = 20, draw.graph, output.errors, DATA_path,
  exclude.interval = NULL)
}
\arguments{
\item{site}{provide the full name of your site (in lower case) i.e. 'barro colorado island'}

\item{stem}{TRUE or FALSE, are you using the stem data (stem=TRUE) or tree data (i.e. called 'full')?}

\item{taper.correction}{TRUE or FALSE, are you willing to apply Cushman et al (2014) taper correction?}

\item{fill.missing}{TRUE or FALSE, are you willing to extrapolate missing DBH from surrounding DBH?}

\item{palm}{TRUE or FALSE, if TRUE, biomass of palm trees is computed through a specific allometric model (Goodman et al. 2013)}

\item{strangler}{TRUE or FALSE, if TRUE, strangler figs tree are flagged (upon a list to published soon)}

\item{maxrel}{a numeric value setting the threshold over which relative productivity is assumed to be too high (usually set at 20 percents)}

\item{draw.graph}{TRUE or FALSE, if TRUE, draw graph for trees with 'erroneous' DBH measures/productivity}

\item{output.errors}{TRUE or FALSE, if TRUE, creat a CSV file with all trees with erroneous DBH measures/productivity}

\item{DATA_path}{allows to provide a different path where the data are located}

\item{exclude.interval}{NULL by default. If needed a vector (e.g. c(1,2)) indicating which census interval(s) must be discarded from computation due, for instance, to a change in  protocol of measurment}
}
\value{
a data.table (data.frame) with all relevant variables.
}
\description{
Main routine to format and correct CTFS-formated data
}
\author{
Ervan Rutishauser (ervan.rutishauser@infoflora.ch)
}
