% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{flag.errors}
\alias{flag.errors}
\title{Flag major errors}
\usage{
flag.errors(DF, site, strangler, maxrel, draw.graph, output.errors,
  exclude.interval)
}
\arguments{
\item{DF}{a data.table}

\item{site}{provide the full name of your site (in lower case) i.e. 'barro colorado island'}

\item{strangler}{TRUE or FALSE, if TRUE, strangler figs tree are flagged (upon a list to published soon)}

\item{maxrel}{a numeric value setting the threshold over which relative productivity is assumed to be too high (usually set at 20 percents)}

\item{draw.graph}{TRUE or FALSE, if TRUE, draw graph for trees with 'erroneous' DBH measures/productivity}

\item{output.errors}{TRUE or FALSE, if TRUE, creat a CSV file with all trees with erroneous DBH measures/productivity}

\item{exclude.interval}{a vector (i.e. c(1,2)) indicating if a set of census intervals must be discarded from computation due for instance to a change in  protocol of measurment}
}
\value{
a data.table (data.frame) with all relevant variables.
}
\description{
Identify trees with major errors in DBH measurments. A major error correspond to a relative individal productivity (or growth) is above a given percentage (set by 'maxrel') of the mean productivity computed at a site. Additionnaly, flagged trees that died at next census interval are also flagged. Option to see DBH measurement (=draw.graph) of flagged trees or print a csv (output.errors) are given.
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
