% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{format.interval}
\alias{format.interval}
\title{Format census intervals}
\usage{
\method{format}{interval}(df, strangler)
}
\arguments{
\item{df}{a data.table}

\item{strangler}{TRUE or FALSE, if TRUE, strangler figs tree are flagged (upon a list to published soon)}
}
\value{
a formated data.table.
}
\description{
Create census intervals (i.e. put consecutive census side by side), assign status by tree (i.e. alive (A),dead (D), recruited (R) or resprout (Rsp))
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
