% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{loess.function}
\alias{loess.function}
\title{CTFS-formated data preparation}
\usage{
loess.function(x, var, range)
}
\arguments{
\item{x}{a data.table}

\item{var}{the name of the variable to be smoothed again intial AGB}

\item{range}{the range of initial AGB to be used for prediction (i.e. 5th and 95th percentiles of the whole distribution)}

\item{site}{provide the full name of your site (in lower case) i.e. 'barro colorado island'}

\item{stem}{TRUE or FALSE, are you using the stem data (stem=TRUE) or tree data (i.e. called 'full')?}

\item{taper.correction}{TRUE or FALSE, are you willing to apply Cushman et al (2014) taper correction?}

\item{fill_missing}{TRUE or FALSE, are you willing to extrapolate missing DBH from surrounding DBH?}

\item{palm}{TRUE or FALSE, if TRUE, biomass of palm trees is computed through a specific allometric model (Goodman et al. 2013)}

\item{strangler}{TRUE or FALSE, if TRUE, strangler figs tree are flagged (upon a list to published soon)}

\item{maxrel}{a numeric value setting the threshold over which relative productivity is assumed to be too high (usually set at 20 percents)}

\item{draw.graph}{TRUE or FALSE, if TRUE, draw graph for trees with 'erroneous' DBH measures/productivity}

\item{output.errors}{TRUE or FALSE, if TRUE, creat a CSV file with all trees with erroneous DBH measures/productivity}

\item{DATA_path}{allows to provide a different path where the data are located}

\item{exclude.interval}{a vector (i.e. c(1,2)) indicating if a set of census intervals must be discarded from computation due for instance to a change in  protocol of measurment}
}
\value{
a data.table (data.frame) with all relevant variables.

a smoothed prediction of the variable of interest
}
\description{
Main routine to format and correct CTFS-formated data

a wrapper to get smoothed predictions of AGB fluxes using a loess function (library 'locfit')
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)

Ervan Rutishauser (er.rutishauser@gmail.com)
}
