% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBfluxes_functions.R
\name{normal.stat}
\alias{normal.stat}
\title{Normalized tree status}
\usage{
normal.stat(X)
}
\arguments{
\item{x}{a data.table}
}
\value{
a data.table (data.frame) with all relevant variables.
}
\description{
Check the consistency of stem/tree status over time (i.e. a tree that is 'alive' at last census can not be 'dead' inbetween)
}
\author{
Ervan Rutishauser (er.rutishauser@gmail.com)
}
